backend/requirements.txt
fastapi
uvicorn
huggingface_hub
chromadb
sentence-transformers
torch
python-dotenv
pandas
requests
firebase-admin
sqlalchemy
psycopg2-binary
